# specs/03-api/openapi.yaml
openapi: 3.0.3
...
info:
  title: Identity Hub API
  version: 0.1.0
  description: |
    Proveedor de identidad: alta administrativa, autenticación, emisión y rotación de
    tokens, segundo factor, autorización de aplicaciones, control por roles y auditoría.

    Cada operación declara `x-requirement` con el identificador del requisito que
    implementa (ver specs/01-requirements.md).

    El refresh token viaja exclusivamente en la cookie `refresh_token`, con
    `HttpOnly`, `Secure`, `SameSite=Strict` y `Path=/api/v1/auth`. Las operaciones
    de refresh y logout exigen esa cookie y usan `POST`; la renovación devuelve JSON
    y logout confirma con `204`. El resto de la API usa exclusivamente
    `Authorization: Bearer`, nunca cookies, por lo que no queda expuesta a CSRF.

    La API de la consola se entrega a través de Nginx y comparte origen. CORS solo se
    habilita para el origen configurado de Contabilidad en `POST /oauth/token` y en el
    JWKS, siempre sin credenciales. Los nombres OAuth en `snake_case` son la única
    excepción intencional a la convención `camelCase`: los define RFC 6749.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Desarrollo local (a través de Nginx)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Solicitar un enlace de restablecimiento
      description: Responde `202` exista o no la cuenta, para no revelar qué correos están registrados.
      x-requirement: RF-015
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailRequest' }
      responses:
        '202': { description: 'Si la cuenta existe, se envió el correo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: El servicio de restablecimiento no está disponible
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags: [auth]
      summary: Fijar una contraseña nueva con el token recibido
      x-requirement: RF-015
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetConfirmRequest' }
      responses:
        '204': { description: Contraseña actualizada y sesiones revocadas; el aviso por correo se publica tras confirmar y un fallo del broker no cambia el resultado (D16) }
        '400': { $ref: '#/components/responses/BadRequest' }
        '410':
          description: El token expiró o ya fue usado
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ── OAuth 2.0 para aplicaciones cliente ──────────────────────────────────
  /oauth/authorize:
    get:
      operationId: authorizeClient
      tags: [auth]
      summary: Autorizar al cliente mediante Authorization Code con PKCE
      description: |
        Valida el cliente configurado, la coincidencia exacta de `redirect_uri`, `state`
        obligatorio y PKCE S256. Si existe la cookie SSO del Hub (`HttpOnly`, `Secure`,
        `SameSite=Lax`) no vuelve a pedir contraseña. Sin sesión SSO
        válida responde `302` al login del Hub con `continue` = la ruta relativa de esta misma
        petición; el login solo acepta `continue` si es una ruta relativa que empieza por
        `/oauth/authorize` (nunca una URL absoluta ni `//host`, para evitar redirección abierta)
        y vuelve a ella tras completar el MFA. Los errores con una URI validada se
        devuelven por redirección según RFC 6749 §4.1.2.1. No se habilita CORS aquí.
        Los parámetros en `snake_case` son nombres de protocolo definidos por RFC 6749.
      x-requirement: RF-020
      security: []
      parameters:
        - name: client_id
          in: query
          required: true
          schema: { type: string }
        - name: redirect_uri
          in: query
          required: true
          schema: { type: string, format: uri }
        - name: response_type
          in: query
          required: true
          schema: { type: string, enum: [code] }
        - name: state
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: code_challenge
          in: query
          required: true
          schema: { type: string, minLength: 43, maxLength: 128 }
        - name: code_challenge_method
          in: query
          required: true
          schema: { type: string, enum: [S256] }
      responses:
        '302':
          description: Redirección al cliente con `code` y `state`, o con `error` y el mismo `state`.
          headers:
            Location:
              required: true
              schema: { type: string, format: uri }
        '400':
          description: Petición inválida que no puede redirigirse de forma segura.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /oauth/token:
    post:
      operationId: exchangeAuthorizationCode
      tags: [auth]
      summary: Canjear un código de autorización por un access token
      description: |
        Acepta exclusivamente `application/x-www-form-urlencoded`. CORS se permite sin
        credenciales y solo para el origen configurado del cliente. No emite refresh token.
        Los campos en `snake_case` son nombres de protocolo definidos por RFC 6749.
      x-requirement: RF-020
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/OAuthTokenRequest' }
      responses:
        '200':
          description: Access token dirigido al cliente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthTokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }

  # ── Titular ───────────────────────────────────────────────────────────────
...
  /api/v1/admin/users/{userId}/invitation:
    post:
      operationId: resendInvitation
      tags: [admin]
      summary: Reenviar la invitación de una cuenta pendiente
      description: Anula invitaciones anteriores y envía un enlace nuevo que vence en 24 horas.
      x-requirement: RF-001
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204': { description: Invitación reenviada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: La cuenta no está pendiente de verificación
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token Ed25519. Validar contra el JWKS; se rechaza cualquier `alg` distinto de `EdDSA`.
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Petición mal formada o que no supera la validación
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Credenciales ausentes, inválidas o expiradas
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Autenticado pero sin permiso para este recurso
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: El recurso no existe
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Límite de peticiones excedido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: Error en formato RFC 7807. Nunca revela si un correo está registrado.
      required: [type, title, status]
      properties:
        type:     { type: string, format: uri, example: 'https://identity.local/problems/invalid-credentials' }
        title:    { type: string, example: 'Credenciales inválidas' }
        status:   { type: integer, example: 401 }
        detail:   { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Detalle por campo en errores de validación
          items:
            type: object
            required: [field, message]
            properties:
              field:   { type: string }
              message: { type: string }
...
    EmailRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetConfirmRequest:
      type: object
      required: [token, password]
      properties:
        token:    { type: string }
        password: { type: string, minLength: 12, maxLength: 128, format: password }
...
    OAuthTokenRequest:
      type: object
      required: [grant_type, code, redirect_uri, client_id, code_verifier]
      properties:
        grant_type:   { type: string, enum: [authorization_code] }
        code:         { type: string }
        redirect_uri: { type: string, format: uri }
        client_id:    { type: string }
        code_verifier: { type: string, minLength: 43, maxLength: 128 }
    OAuthTokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token: { type: string, description: JWT Ed25519 con `aud` del cliente y solo sus roles }
        token_type:   { type: string, enum: [Bearer] }
        expires_in:   { type: integer, example: 900 }
...